+++ library/core/src/main/java/com/google/android/exoplayer2/mediacodec/MediaCodecRenderer.java
@@ -1342,7 +1342,7 @@ public abstract class MediaCodecRenderer extends BaseRenderer {
     resetInputBuffer();
     codecReceivedBuffers = true;
     codecReconfigurationState = RECONFIGURATION_STATE_NONE;
-    decoderCounters.inputBufferCount++;
+    decoderCounters.queuedInputBufferCount++;
     return true;
   }