+++ library/core/src/main/java/com/google/android/exoplayer2/video/DecoderVideoRenderer.java
@@ -757,7 +757,7 @@ public abstract class DecoderVideoRenderer extends BaseRenderer {
         decoder.queueInputBuffer(inputBuffer);
         buffersInCodecCount++;
         decoderReceivedBuffers = true;
-        decoderCounters.inputBufferCount++;
+        decoderCounters.queuedInputBufferCount++;
         inputBuffer = null;
         return true;
       default: